requirements.txt
attrs==20.3.0
autopep8==1.5.5
beautifulsoup4==4.11.1
blinker==1.4
cachelib==0.9.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==6.4.2
elastic-transport==8.1.2
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Session==0.4.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.12.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.5
Werkzeug==1.0.1
